Add interface StrictMap, StrictList and StrictSet. Renaming existing implementations and change usages to use the interfaces where possible
