+++ src/main/java/org/mineacademy/fo/collection/AutoUpdateMap.java
@@ -18,7 +18,7 @@ public final class AutoUpdateMap<V, K> extends AbstractStrictCollection {
 	/**
 	 * The internal map
 	 */
-	private final StrictMap<V, K> map = new StrictMap<>();
+	private final StrictLinkedHashMap<V, K> map = new StrictLinkedHashMap<>();
 
 	/**
 	 * The code that gets triggered
@@ -59,7 +59,7 @@ public final class AutoUpdateMap<V, K> extends AbstractStrictCollection {
 	 *
 	 * @return
 	 */
-	public StrictMap<V, K> getSource() {
+	public StrictMap<V,K> getSource() {
 		return map;
 	}
